# api.yaml
openapi: 3.0.3
info:
  title: API for project 3d-inventory
  description: >-
    REST API for **3d-inventory**. Project is a simple solution that allows you to build a spatial and database
    representation of all types of warehouses and server rooms.
  contact:
    name: 3d-inventory-mongo-api
    url: https://github.com/karol-preiskorn/3d-inventory-mongo-api/
  version: 0.54.0
servers:
  - url: https://127.0.0.1:8080
  - url: http://0.0.0.0:8080
  - url: https://0.0.0.0:8080
  - url: http://localhost:8080
  - url: https://localhost:8080
  - url: http://127.0.0.1:8080
  - url: http://172.17.0.2:8080
  - url: http://172.17.0.3:8080
  - url: http://172.20.0.3:8080
  - url: https://d-inventory-api-wzwe3odv7q-ew.a.run.app
  - url: https://3d-inventory-api.ultimasolution.pl
  - url: https://3d-inventory.ultimasolution.pl
...
paths:
  /login:
    post:
      tags:
        - authentication
      summary: User login to obtain JWT token
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
              required:
                - username
      responses:
        '200':
          description: JWT token issued
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
        '400':
          description: Invalid credentials
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '404':
          description: Not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
...
  /attributes:
    get:
      tags:
        - attributes
      summary: Get all attributes
      description: >-
        Retrieve a list of all attributes. You can apply optional query parameters such as `key` to filter
        attributes by their key or `type` to filter by their type.
      operationId: getAllAttributes
      security:
        - ViewerAuth: []
      responses:
        '200':
          description: A list of attributes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Attributes'
        '401':
          description: Unauthorized - Missing or invalid authentication
        '403':
          description: Forbidden - Insufficient privileges (viewer role or higher required)
        '404':
          description: Not found.
        '406':
          description: Not Acceptable
        '429':
          description: Too Many Requests
        '500':
          description: Unexpected server error
        default:
          description: Unexpected error
      x-swagger-router-controller: Attributes
    put:
      tags:
        - attributes
      summary: Update an existing attribute
      description: Update the details of an existing attribute.
      operationId: updateAttribute
      security:
        - UserAuth: []
      parameters:
        - name: id
          in: path
          description: ID of the attribute to update
          required: true
          style: simple
          explode: false
          schema:
            type: string
            example: 6566ed4ace63d42b25fbba19
            pattern: ^[0-9A-Fa-f]{24}$
            minLength: 24
            maxLength: 24
            x-42c-format: o:uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Attribute'
        required: true
      responses:
        '200':
          description: Attribute updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Attribute'
        '401':
          description: Unauthorized - Missing or invalid authentication
        '403':
          description: Forbidden - Insufficient privileges (user role or higher required)
        '404':
          description: Attribute update data is inconsistent.
        '406':
          description: Not Acceptable
        '429':
          description: Too Many Requests
        '500':
          description: Unexpected server error
        default:
          description: Unexpected error
      x-swagger-router-controller: Attributes
    post:
      tags:
        - attributes
      summary: Create a new attribute
      description: Create a new attribute. Provide attribute details in the request body.
      operationId: createAttribute
      security:
        - UserAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Attribute'
      responses:
        '200':
          description: Attribute created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Attribute'
        '400':
          description: Bad request.
        '401':
          description: Unauthorized - Missing or invalid authentication
        '403':
          description: Forbidden - Insufficient privileges (user role or higher required)
        '500':
          description: Unexpected server error
      x-swagger-router-controller: Attributes
    delete:
      tags:
        - attributes
      summary: Delete all attributes
      description: Remove all attributes from the inventory.
      operationId: deleteAllAttributes
      security:
        - AdminAuth: []
      responses:
        '200':
          description: All attributes deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Confirmation'
        '401':
          description: Unauthorized - Missing or invalid authentication
        '403':
          description: Forbidden - Insufficient privileges (admin role required)
        '404':
          description: No attributes found to delete
      x-swagger-router-controller: Attributes
...
  /devices:
    get:
      tags:
        - devices
      summary: Get all devices
      description: Retrieve a list of all devices.
      operationId: getAllDevices
      security:
        - ViewerAuth: []
      responses:
        '200':
          description: A list of devices
        '401':
          description: Unauthorized - Missing or invalid authentication
        '403':
          description: Forbidden - Insufficient privileges (viewer role or higher required)
        '404':
          description: No devices found.
        '406':
          description: Not Acceptable
        '429':
          description: Too Many Requests
        '500':
          description: Unexpected server error
        default:
          description: Unexpected error
      x-swagger-router-controller: Devices
    put:
      tags:
        - devices
      summary: Update an existing device
      description: Update the details of an existing device.
      operationId: updateDevice
      security:
        - UserAuth: []
      responses:
        '200':
          description: Device updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
        '401':
          description: Unauthorized - Missing or invalid authentication
        '403':
          description: Forbidden - Insufficient privileges (user role or higher required)
        '404':
          description: Device update data is inconsistent.
        '406':
          description: Not Acceptable
        '429':
          description: Too Many Requests
        '500':
          description: Unexpected server error
        default:
          description: Unexpected error
      x-swagger-router-controller: Devices
    delete:
      tags:
        - devices
      summary: Delete all devices
      description: Remove all devices from the inventory.
      operationId: deleteAllDevices
      security:
        - AdminAuth: []
      responses:
        '200':
          description: All devices deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Confirmation'
        '401':
          description: Unauthorized - Missing or invalid authentication
        '403':
          description: Forbidden - Insufficient privileges (admin role required)
        '404':
          description: No devices found to delete
      x-swagger-router-controller: Devices
...
  /logs:
    get:
      tags:
        - logs
      summary: Get logs from database
      description: Get all logs.
      operationId: logsGET
      security:
        - ViewerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Logs'
        '401':
          description: Unauthorized - Missing or invalid authentication
        '403':
          description: Forbidden - Insufficient privileges (viewer role or higher required)
        '404':
          description: No found
        '500':
          description: Error
      x-swagger-router-controller: Logs
    put:
      tags:
        - logs
      summary: Update all set log
      description: Update entire collection logs.
      operationId: logPUT
      security:
        - AdminAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                _id:
                  type: string
                  example: 6566ed4ace63d42b25fbba19
                  pattern: ^[0-9A-Fa-f]{24}$
                  minLength: 24
                  maxLength: 24
                  x-42c-format: o:uuid
                object:
                  type: string
                  example: 6566ed4ace63d42b25fbba19
                  pattern: ^[0-9A-Fa-f]{24}$
                  minLength: 24
                  maxLength: 24
                  x-42c-format: o:uuid
                date:
                  type: string
                  description: Date as string
                operation:
                  type: string
                  description: CRUD CREATE/UPDATE/DELETE
                component:
                  type: string
                  description: Device, Connection, Model, Floor
                message:
                  type: object
                  description: Log message
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Log'
        '401':
          description: Unauthorized - Missing or invalid authentication
        '403':
          description: Forbidden - Insufficient privileges (admin role required)
        '404':
          description: Id Not found
      x-swagger-router-controller: Logs
    post:
      tags:
        - logs
      summary: Insert new log
      description: Create log.
      operationId: logPOST
      security:
        - UserAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Log'
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Log'
              example:
                $ref: '#/components/schemas/Log/example'
        '401':
          description: Unauthorized - Missing or invalid authentication
        '403':
          description: Forbidden - Insufficient privileges (user role or higher required)
        '404':
          description: Error log not found
        '500':
          description: Error log not created
      x-swagger-router-controller: Logs
    delete:
      tags:
        - logs
      summary: Delete log
      description: Delete all logs or a specific log by ID.
      operationId: logAllDelete
      security:
        - AdminAuth: []
      responses:
        '200':
          description: Success log deleted
        '401':
          description: Unauthorized - Missing or invalid authentication
        '403':
          description: Forbidden - Insufficient privileges (admin role required)
        '404':
          description: Error log not found
      x-swagger-router-controller: Logs
...
  /connections:
    get:
      tags:
        - connections
      summary: >-
        Server heartbeat operation. Get information about connection between Devices.
      description: Get all connections between devices
      operationId: connectionsGet
      security:
        - ViewerAuth: []
      responses:
        '200':
          description: Success get connection between devices
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Connections'
        '401':
          description: Unauthorized - Missing or invalid authentication
        '403':
          description: Forbidden - Insufficient privileges (viewer role or higher required)
      x-swagger-router-controller: Connections
    put:
      tags:
        - connections
      summary: Update new connection between two devices
      description: >-
        Create connection between two devices. Two id and attributes from attributes.
      operationId: connectionPUT
      security:
        - UserAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Connections'
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Connections'
              example:
                $ref: '#/components/schemas/Connections/example'
        '401':
          description: Unauthorized - Missing or invalid authentication
        '403':
          description: Forbidden - Insufficient privileges (user role or higher required)
        '404':
          description: Not found connection data
        '500':
          description: Error
      x-swagger-router-controller: Connections
    post:
      tags:
        - connections
      summary: Create new connection between two devices
      description: >-
        Create connection between two devices. Two id and attributes from attributes.
      operationId: connectionPOST
      security:
        - UserAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Connections'
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Connection'
              example:
                $ref: '#/components/schemas/Connection/example'
        '401':
          description: Unauthorized - Missing or invalid authentication
        '403':
          description: Forbidden - Insufficient privileges (user role or higher required)
        '404':
          description: Not found connection data
        '500':
          description: Error
      x-swagger-router-controller: Connections
    delete:
      tags:
        - connections
      summary: Delete connection
      description: >-
        Delete connection between two devices. Two id and attributes from attributes.
      operationId: connectionDELETE
      security:
        - AdminAuth: []
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Confirmation'
        '401':
          description: Unauthorized - Missing or invalid authentication
        '403':
          description: Forbidden - Insufficient privileges (admin role required)
        '404':
          description: Not found connection to delete
        '500':
          description: Error
      x-swagger-router-controller: Connections
  /connections/devices:
    get:
      tags:
        - connections
      summary: Get information about one Connection and its devices.
      description: Get all list Connection with information about devices and their attributes
      operationId: connectionGetAndDevices
      responses:
        '200':
          description: Success get connection between devices
        '404':
          description: Not found connection data
        '500':
          description: Error
      x-swagger-router-controller: Connections
...
  /attributesDictionary:
    get:
      tags:
        - attributesDictionary
      summary: Information about configuration attributes types.
      description: >-
        Store information about what parameters can apply to what type of type objects.
      operationId: attributesDictionaryGet
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttributesDictionaries'
        '400':
          description: Bad request.
        '404':
          description: Not found.
        '500':
          description: Unexpected server error.
      x-swagger-router-controller: attributesDictionary
    put:
      tags:
        - attributesDictionary
      summary: Update attributesDictionary
      operationId: attributesDictionaryPut
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AttributesDictionary'
      responses:
        '200':
          description: OK. Indicates that the update request has succeeded.
        '400':
          description: Bad request.
        '404':
          description: Not found.
        '500':
          description: >-
            Unexpected server error. The server encountered an unexpected condition that prevented it
            fromfulfilling the request.
      x-swagger-router-controller: attributesDictionary
    post:
      tags:
        - attributesDictionary
      summary: Create devices attributes types.
      description: 3d-inventory attributes from dictionary attributes
      operationId: attributesDictionaryPost
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AttributesDictionary'
      responses:
        '200':
          description: OK. Indicates that the request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttributesDictionary'
              example:
                $ref: '#/components/schemas/AttributesDictionary/example'
        '400':
          description: Bad request.
        '500':
          description: >-
            Unexpected server error. The server encountered an unexpected condition that prevented it
            fromfulfilling the request.
      x-swagger-router-controller: attributesDictionary
    delete:
      tags:
        - attributesDictionary
      summary: Delete all attributesDictionary
      description: |-
        Delete a specific attribute in the attributes dictionary by ID. '200':
          description: OK. Indicates that the request has succeeded.
      responses:
        '200':
          description: OK. Indicates that the request has succeeded.
        '404':
          description: Not found
        '500':
          description: >-
            Unexpected server error. The server encountered an unexpected condition that prevented it
            fromfulfilling the request.
      x-swagger-router-controller: attributesDictionary
...
  /attributesDictionary/models:
    get:
      tags:
        - attributesDictionary
      summary: Get information about devices attributes for models.
      description: Dictionary attributes dictionary for models.
      operationId: attributesDictionaryModelsGet
      responses:
        '200':
          description: OK. Indicates that the request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttributesDictionaries'
              example:
                $ref: '#/components/schemas/AttributesDictionaries'
        '400':
          description: Bad request.
        '500':
          description: >-
            Unexpected server error. The server encountered an unexpected condition that prevented it
            fromfulfilling the request.
      x-swagger-router-controller: attributesDictionary
  /attributesDictionary/devices:
    get:
      tags:
        - attributesDictionary
      summary: Get information about devices attributes for models.
      description: Dictionary attributes dictionary for models.
      operationId: attributesDictionaryDevicesGet
      responses:
        '200':
          description: OK. Indicates that the request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttributesDictionaries'
              example:
                $ref: '#/components/schemas/AttributesDictionaries'
        '400':
          description: Bad request.
        '500':
          description: >-
            Unexpected server error. The server encountered an unexpected condition that prevented it
            from  fulfilling the request.
      x-swagger-router-controller: attributesDictionary
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer token for authentication. Include role and permissions for authorization.
    AdminAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer token with admin role required
    UserAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer token with user role or higher required
    ViewerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer token with viewer role or higher required (read-only access)
  schemas:
...
    AttributesDictionaries:
      type: array
      description: List of Attribute Dictionary
      maxItems: 1000
      items:
        $ref: '#/components/schemas/AttributesDictionary'
...
    Confirmation:
      type: object
      properties:
        message:
          type: string
      additionalProperties: false
      description: Universal confirmation message from database
      example:
        message: Succeed
...